Move Launcher into root folder and remove redundant code in seeyes
