+++ src/main/java/seeyes/Launcher.java
@@ -1,7 +1,6 @@
-package seeyes.ui;
+package seeyes;
 
 import javafx.application.Application;
-import seeyes.Main;
 
 /**
  * A launcher class to workaround classpath issues.
