+++ src/main/java/seeyes/Seeyes.java
@@ -1,17 +1,13 @@
 package seeyes;
 
-import java.util.Scanner;
-
 import javafx.animation.PauseTransition;
 import javafx.util.Duration;
 import seeyes.command.Command;
 import seeyes.command.CommandResult;
 import seeyes.exception.InvalidCommandException;
-import seeyes.exception.NoMoreCommandsException;
 import seeyes.exception.StorageException;
 import seeyes.parser.Parser;
 import seeyes.storage.Storage;
-import seeyes.task.Task;
 import seeyes.task.TaskList;
 import seeyes.ui.Ui;
 
@@ -21,7 +17,6 @@ import seeyes.ui.Ui;
 public class Seeyes {
     private TaskList taskList;
     private Storage storage;
-    private Scanner scanner;
     private Ui ui;
 
     /**
@@ -31,7 +26,6 @@ public class Seeyes {
      *            the path to the data file
      */
     public Seeyes(String filePath) {
-        scanner = new Scanner(System.in);
         storage = new Storage(filePath, taskList);
         ui = Ui.getUi();
 
@@ -46,13 +40,6 @@ public class Seeyes {
     }
 
     /**
-     * Prints the current size of the task list.
-     */
-    public void printListSize() {
-        ui.say("Number of tasks in list: " + taskList.size());
-    }
-
-    /**
      * Sets the task list.
      *
      * @param taskList
@@ -61,73 +48,6 @@ public class Seeyes {
     public void setTaskList(TaskList taskList) {
         assert taskList instanceof TaskList : "taskList should be a TaskList instance.";
         this.taskList = taskList;
-    }
-
-    /**
-     * Prints the current task list.
-     */
-    public void printList() {
-        assert taskList instanceof TaskList : "taskList should be a TaskList instance.";
-        CommandResult result;
-        if (taskList.size() == 0) {
-            result = new CommandResult(
-                    "list is empty! add your first item with 'todo [item]'.");
-            return;
-        } else {
-            result = new CommandResult(
-                    "You have " + taskList.size() + " items in your list.",
-                    taskList.getTaskList());
-        }
-        ui.showResult(result);
-    }
-
-    /**
-     * Adds a task to the list and displays the result.
-     *
-     * @param task
-     *            the task to add
-     */
-    public void addToList(Task task) {
-        assert task instanceof Task : "Task should be a Task class instance.";
-        CommandResult result;
-        if (taskList.addTask(task)) {
-            result = new CommandResult("Added: " + task.toString());
-        } else {
-            result = new CommandResult("Failed to add: " + task.toString());
-        }
-        ui.showResult(result);
-        printListSize();
-    }
-
-    /**
-     * Runs the main application loop.
-     */
-    public void run() {
-        while (true) {
-            try {
-                Command command = Parser.parseUserInput(ui.getNextUserCommand())
-                        .setData(taskList, storage);
-                CommandResult result = command.execute();
-                if (result.getTaskList().isPresent()) {
-                    taskList = result.getTaskList().get();
-                }
-                ui.showResult(result);
-
-                // Solution below inspired by https://github.com/donkoo24/ip/blob/master/src/main/java/lux/Lux.java
-                if (command.isExit()) {
-                    break;
-                }
-            } catch (InvalidCommandException e) {
-                ui.showError(e.getMessage());
-            } catch (NoMoreCommandsException e) {
-                ui.showError(e.getMessage());
-                break;
-            }
-        }
-
-        // Terminate
-        scanner.close();
-        exit();
     }
 
     /**
